requirements.txt
requests
fastapi
uvicorn[standard]
python-decouple==3.8
gunicorn
openai
bcrypt==4.0.1
passlib==1.7.4
pydantic
pymupdf4llm
mammoth
pandas
XlsxWriter
pyarrow